i965/bxt: Add environment override for thread count
Because of the uncertainty of silicon fusing at power on, I've decided
to modify the MESA driver to accept a user environment variable to hint
to the driver what thread count to use that would depend on whether the
SKU being operated on has the POR 2x3 subslice or the capped 2x2. To get
the more, the user would need to "export MESA_965_BXT_FAST=anything".
Else unset that environment.

Signed-off-by: Alan Previn Teres Alexis <alan.previn.teres.alexis@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>