configure.ac: enable_asm=yes when x-compiling across same arch (v2)
Currently, configure script forces to disable asm when
cross-compilation performed. This is based on an assumption
that host architecure and target architecture are different.
But there's always a case that host and target can share the
same ASM codes when both architures are same or compatible,
which the current script can't cover.

This change addresses the limitation by reseting enable_asm
only if two CPUs (host and target) are different.

v2: removed unrelated changes

Signed-off-by: Dongwon Kim <dongwon.kim@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>