egl: EGL_KHR_reusable_sync implementation (v5)
This patch adds an EGL extension, EGL_KHR_reusable_sync,
that is basically adding a way for multiple APIs or threads
to be synchronized to each other in terms of execution order.

All the functionalities described in the specificiation
of EGL_KHR_reusable_sync have been implemented on EGL
layer and all of additions are independent from driver layer.
This means enabling this extension shouldn't be limited
to any specific HW.

An official specification doc can be found at

https://www.khronos.org/registry/egl/extensions/KHR/EGL_KHR_reusable_sync.txt

v2: removed 'type' from dri2_egl_sync because of redundancy
    ((_EGL_sync)base.Type can be used instead),
    added reusable sync support to eglGetSyncAttribKHR
    null-check on *value in eglGetSyncAttribKHR

v3: removed 'reusable sync' type from the case when
    attribute is "EGL_SYNC_CONDITION_KHR" in eglGetSyncAttribKHR
    because this attribute is not supported.

v4: fixed a deadlock situation in eglClientWaitSyncKHR in case
    more than 1 thread that reference same EGLDisplay are trying
    to get in blocked (waiting) state via this API.

v5: fixed a race condition in dri2_destroy_sync in case of
    "EGL_SYNC_REUSABLE_KHR"
    no reason to protect pthread_cond_broadcast with mutex,
    so removed

Signed-off-by: Dongwon Kim <dongwon.kim@intel.com>
Reviewed-by: James Xiong <james.xiong@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>